% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDL_cropscape.R
\name{download_state_data_parallel}
\alias{download_state_data_parallel}
\title{Download State Data in Parallel}
\usage{
download_state_data_parallel(State, year)
}
\arguments{
\item{State}{A character string specifying the state for which data is to be downloaded.}

\item{year}{An integer specifying the year for which data is to be downloaded.}
}
\value{
A list containing the downloaded CDL data for each county in the specified state.
Each element of the list is named by the county's FIPS code and contains the data as returned
by the CropScapeR::GetCDLComp function.
}
\description{
Downloads Cropland Data Layer (CDL) composited data for a specified state and year
using CropScapeR::GetCDLComp function. The data for each county within the state is downloaded
in parallel and stored in a list, with each list element named after the county's FIPS code.
}
\examples{
\dontrun{
  alabama_data_2017 <- download_state_data_parallel(State = "AL", year = 2017)
}

}
