% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDL_cropscape.R
\docType{data}
\name{Lookup_table}
\alias{Lookup_table}
\title{Lookup Table for Land Uses}
\format{
A data frame with 136 rows and 2 columns
including columns such as 'Land use Code', 'Land use Description'.
}
\source{
Description of the Land use.
}
\usage{
Lookup_table
}
\description{
A dataset containing the lookup table for all land uses, sourced from "./data-raw/Lookup_table.csv".
This dataset is used to map land use codes to their descriptions.
}
\keyword{datasets}
